requirements.txt
fastapi
uvicorn[standard]
sqlmodel
passlib
python-jose[cryptography]
jinja2
email-validator
itsdangerous
httpx
python-multipart
bcrypt